requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
openai